% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid.R
\name{calculate_centroid}
\alias{calculate_centroid}
\title{calculate_centroid}
\usage{
calculate_centroid(p, chr_info, results_gABC_list)
}
\arguments{
\item{p}{An element of the list to which the centroid information will be added.}

\item{chr_info}{Chromosome information.}

\item{results_gABC_list}{results_gABC_list after valid vector generation.}
}
\description{
This function will calculate centroid of gABC scores for enhancer candidates.
}
