% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_centroid_remerge.R
\name{calculate_centroid_remerge}
\alias{calculate_centroid_remerge}
\title{calculate_centroid_remerge}
\usage{
calculate_centroid_remerge(env_T, x, chr_info, results_gABC_list)
}
\arguments{
\item{env_T}{An element of the list containing co-accessibility network edges split by decomposed component ID after Link Loss Correction.}

\item{x}{Query or Subject decomposed component to calculate centroid of gABC scores.}

\item{chr_info}{Chromosome information.}

\item{results_gABC_list}{results_gABC_list after generating decomposed components.}
}
\description{
This function will calculate centroid of gABC scores for decomposed components to perform Sparsity Correction.
}
