% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_gABC_score.R
\name{Calculate_gABC_score}
\alias{Calculate_gABC_score}
\title{Calculate_gABC_score}
\usage{
Calculate_gABC_score(
  data_lump_enCORE,
  n_col,
  STARE_dir,
  output_dir,
  threshold_score = 0.01
)
}
\arguments{
\item{data_lump_enCORE}{A list containing the data required for running enCORE.}

\item{n_col}{The number of columns in peak-cluster CPM matrix for gABC scoring.}

\item{STARE_dir}{Directory containing STARE_ABCpp.}

\item{output_dir}{Output directory where the gABC scoring results will be saved.}

\item{threshold_score}{Threshold for the gABC score. Default = 0.01.}
}
\description{
This function will perform gABC scoring using STARE.
}
