% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_valid_vector.R
\name{convert_valid_vector}
\alias{convert_valid_vector}
\title{convert_valid_vector}
\usage{
convert_valid_vector(x, y)
}
\arguments{
\item{x}{results_gABC_list element}

\item{y}{total_gene_id element}
}
\description{
This function will convert gABC scores to valid vector to perform enCORE.
}
