% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Determine_TF_weight_threshold.R
\name{Determine_TF_weight_threshold}
\alias{Determine_TF_weight_threshold}
\title{Determine_TF_weight_threshold}
\usage{
Determine_TF_weight_threshold(
  proj_atac,
  data_lump_enCORE,
  list_cluster,
  use_default_thres = TRUE,
  predefined_thres = 0.675
)
}
\arguments{
\item{proj_atac}{An ArchRProject object containing chromVAR CISBP v2 motif deviation.}

\item{data_lump_enCORE}{A list containing the data required for running enCORE.}

\item{list_cluster}{A list containing cell population information to be used for performing enCORE.}

\item{use_default_thres}{If TRUE, use predefined_thres as TF weight threshold. If not, use automatically calculated TF weight threshold.}

\item{predefined_thres}{Pre-defined default threshold. Default = 0.675.}
}
\description{
This function will determine TF weight threshold.
}
