% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_decomposed_components.R
\name{extract_decomposed_components}
\alias{extract_decomposed_components}
\title{extract_decomposed_components}
\usage{
extract_decomposed_components(
  link_per_chr,
  results_gABC_list,
  thres_cos = 0.4,
  max_dist = 5e+05,
  n_core = 4
)
}
\arguments{
\item{link_per_chr}{A list containing co-accessibility network edges split by chromosome information.}

\item{results_gABC_list}{results_gABC_list after assigning cosine similarity to co-accessibility network edge.}

\item{thres_cos}{Threshold for the cosine similarity. Default = 0.4.}

\item{max_dist}{Maximum effective distance between enhancer candidates. Default = 500000.}

\item{n_core}{An integer specifying the number of CPU-cores to perform enCORE. Default = 4.}
}
\description{
This function will execute Link Loss Correction.
}
