% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_initial_enhancer_candidates.R
\name{Extract_initial_enhancer_candidates}
\alias{Extract_initial_enhancer_candidates}
\title{Extract_initial_enhancer_candidates}
\usage{
Extract_initial_enhancer_candidates(
  proj_atac,
  data_coacc,
  output_dir,
  organism = "hg38",
  n_core = 4
)
}
\arguments{
\item{proj_atac}{An ArchRProject object.}

\item{data_coacc}{Co-accessibility information from ArchR::getCoAccessibility function.}

\item{output_dir}{Output directory where the required files for gABC scoring will be saved.}

\item{organism}{Organism information. Accepted values are "hg38", "hg19", and "mm10". Default = "hg38".}

\item{n_core}{An integer specifying the number of CPU-cores to perform enCORE. Default = 4.}
}
\description{
This function will extract initial enhancer candidates for gABC scoring.
}
