% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ranges_for_bed.R
\name{extract_ranges_for_bed}
\alias{extract_ranges_for_bed}
\title{extract_ranges_for_bed}
\usage{
extract_ranges_for_bed(x, min_count = 5)
}
\arguments{
\item{x}{Each raw CORE region in test_list variable.}

\item{min_count}{The minimum CPM count for CORE constituents. Only constituents with a CPM above this threshold will be retained.}
}
\description{
This function will convert the extracted CORE regions to .bed file.
}
