% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw_CORE_per_cluster.R
\name{extract_raw_CORE_per_cluster}
\alias{extract_raw_CORE_per_cluster}
\title{extract_raw_CORE_per_cluster}
\usage{
extract_raw_CORE_per_cluster(
  proj_atac,
  data_lump_enCORE,
  n_celltype,
  thres_cos_initial = 0.4,
  thres_cos_link_loss = 0.4,
  thres_cos_re_merge = 0.92,
  thres_corr_add = 0.49,
  thres_topic_add = 0.5,
  max_dist = 5e+05,
  max_nCORE = 0.1 * 26000/1.31,
  max_iter_elbow = 10,
  min_count = 5
)
}
\arguments{
\item{proj_atac}{An ArchRProject object containing chromVAR CISBP v2 motif deviation.}

\item{data_lump_enCORE}{A list containing the data required for running enCORE.}

\item{n_celltype}{Cell population information in list_cluster.}

\item{thres_cos_initial}{Threshold for the cosine similarity to extract interaction-shared decomposed components. Default = 0.4.}

\item{thres_cos_link_loss}{Threshold for the cosine similarity to perform Link Loss Correction. Default = 0.4.}

\item{thres_cos_re_merge}{Threshold for the cosine similarity to perform Sparsity Correction. Default = 0.92.}

\item{thres_corr_add}{Threshold for the co-accessibility value to incorporate additional nodes with high accessibility. Default = 0.49.}

\item{thres_topic_add}{Threshold for the accessibility to incorporate additional nodes with high accessibility. Default value is top 0.5\% accessibility of original enhancer candidates. In other words, Default = 0.5.}

\item{max_dist}{Maximum effective distance between enhancer candidates. Default = 500000.}

\item{max_nCORE}{The maximum number of CORE regions. Default = 0.1*26000/1.31.}

\item{max_iter_elbow}{The maximum number of iterations for elbow point determination. Default = 10.}

\item{min_count}{The minimum CPM count for CORE constituents. Only constituents with a CPM above this threshold will be retained.}
}
\description{
This function will extract raw CORE regions from each cell population.
}
