% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_proximal_filtering.R
\name{iterative_proximal_filtering}
\alias{iterative_proximal_filtering}
\title{iterative_proximal_filtering}
\usage{
iterative_proximal_filtering(
  data_lump_enCORE,
  n_celltype,
  thres_inactive = 12.5,
  thres_eprox = 7500
)
}
\arguments{
\item{data_lump_enCORE}{A list containing the data required for running enCORE.}

\item{n_celltype}{Cell population information in list_cluster.}

\item{thres_inactive}{Threshold for the accessibility within promoter to determine inactive genes. Default = 12.5.}

\item{thres_eprox}{Effective distance of inactive proximal enhancers. Default = 7500 bp.}
}
\description{
This function will execute iterative proximal filtering to extract CORE regions from the active option.
}
