% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remerge_component.R
\name{remerge_component}
\alias{remerge_component}
\title{remerge_component}
\usage{
remerge_component(env_T, chr_info, results_gABC_list, thres_cos = 0.92)
}
\arguments{
\item{env_T}{An element of the list containing co-accessibility network edges split by decomposed component ID after Link Loss Correction.}

\item{chr_info}{Chromosome information.}

\item{results_gABC_list}{results_gABC_list after generating decomposed components.}

\item{thres_cos}{Threshold for the cosine similarity. Default = 0.92.}
}
\description{
This function will re-merge decomposed components with high cosine similarity to perform Sparsity Correction.
}
