% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_peak_id.R
\name{split_by_peak_id}
\alias{split_by_peak_id}
\title{split_by_peak_id}
\usage{
split_by_peak_id(x)
}
\arguments{
\item{x}{results_gABC_list element.}
}
\description{
This function will split each element of the results_gABC_list by peak_id.
}
